% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_api_election.R
\name{election_filter}
\alias{election_filter}
\title{DT Filter for UK Elections}
\usage{
election_filter(data, params)
}
\arguments{
\item{data, params}{Standard parameters required for a funcFilter}
}
\description{
Example of REST API implementation of a DT filter for a UK Elections API
}
\details{
Data is fetched from lda.data.parliament.uk/elections.json

The API provides pagination and sorting, which are supported by this custom
DT filter.
}
