% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{query_sqlite}
\alias{query_sqlite}
\title{Sqlite implementation of a query_fun.}
\usage{
query_sqlite(con, params, tbl)
}
\arguments{
\item{con}{DBI database connection}

\item{params}{named list provided by a datatable containing sorting, filtering
and pagination data.}

\item{tbl}{string, the table/view in sqlite.}
}
\description{
Simple implementation showing how to implement a \code{query_fun} for a sqlite DB.
}
\details{
This function can be provided to sql_filter_factory to describe how to
fetch a datatable payload from a sqlite table.

Use as a template to implement other backends.

Will be called via \code{sql_filter_factory} with arguments
\code{con}, \code{page}, and \code{...} where \code{...} are extra arguments
passed to \code{sql_filter_factory}.
}
\examples{
\dontrun{
sql_filter_factory(con, query_sqlite, tbl = "mtcars")
}
}
