% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{query_sqlite}
\alias{query_sqlite}
\title{Sqlite implementation of a query_fun.}
\usage{
query_sqlite(con, params, tbl, id_field = NA)
}
\arguments{
\item{con}{DBI database connection}

\item{params}{named list provided by a datatable containing sorting, filtering
and pagination data.}

\item{tbl}{string, the table/view in sqlite.}

\item{id_field}{string, optional. Field used to identify a row when using \code{input$tbl_rows_selected}.}
}
\description{
Simple implementation showing how to implement a \code{query_fun} for a sqlite DB.
}
\details{
This function can be provided to \code{sql_filter_factory} to describe how to
fetch a datatable payload from a sqlite table.

This implementation provides paging and sorting, but filtering and row indices
filters are not yet implemented.

This function will be called via \code{sql_filter_factory} with arguments

\code{con}, \code{page}, and \code{...} (where \code{...} are any extra
arguments given to \code{sql_filter_factory}). With this example an additional
\code{tbl} parameter has been implemented, which would be expected to be
passed in when \code{sql_filter_factory} is called.
}
\examples{
\dontrun{
sql_filter_factory(con, query_sqlite, tbl = "mtcars")
}
}
